# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AnyLive API
  version: 0.1.0
  description: |
    AnyLive overseas-first live streaming control plane.
    P0 surface: system health + meta. Auth/rooms freeze in v0.1/v0.2.
servers:
  - url: http://localhost:8088
    description: Local API
...
paths:
...
  /api/v1/admin/grant:
    post:
      tags: [admin]
      operationId: adminGrant
      summary: Bootstrap or grant admin role (admin | moderator | ops)
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GrantAdminRequest"
      responses:
        "204": {description: Granted}
  /api/v1/admin/ban:
    post:
      tags: [admin]
      operationId: adminBan
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserRequest"
      responses:
        "204": {description: Banned}
  /api/v1/admin/mute:
    post:
      tags: [admin]
      operationId: adminMute
      summary: Mute a user (blocks chat and gifts)
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MuteUserRequest"
      responses:
        "204":
          description: Muted
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/admin/unmute:
    post:
      tags: [admin]
      operationId: adminUnmute
      summary: Unmute a user
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnmuteUserRequest"
      responses:
        "204":
          description: Unmuted
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/admin/rooms/force-close:
    post:
      tags: [admin]
      operationId: adminForceClose
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Closed}
  /api/v1/admin/audit:
    get:
      tags: [admin]
      operationId: adminAudit
      security: [{bearerAuth: []}]
      responses:
        "200": {description: Audit list}
  /api/v1/admin/wallet/reconcile:
    get:
      tags: [admin]
      operationId: adminWalletReconcile
      summary: Run wallet ledger/balance consistency scan
      description: >
        Compares each user's stored balance to the sum of ledger amounts.
        P1 dogfood gate requires imbalance_count == 0 after gift/topup flows.
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Reconcile report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletReconcileResponse"
  /api/v1/admin/pay/expire-orders:
    post:
      tags: [admin]
      operationId: adminExpirePayOrders
      summary: Expire unpaid pay orders past expires_at
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Expire result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpirePayOrdersResponse"
  /api/v1/admin/analytics/summary:
    get:
      tags: [admin]
      operationId: adminAnalyticsSummary
      summary: Dogfood client-event buffer summary (not warehouse DAU)
      description: >
        Aggregates the in-process analytics ring buffer by event name.
        Use during soft launch to verify client ingest; full DAU/paid boards stay external.
      security: [{bearerAuth: []}]
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummaryResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [code, message, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
...
    BanUserRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
        reason:
          type: string
    GrantAdminRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
        role:
          type: string
          description: admin | moderator | ops (default admin)
          enum: [admin, moderator, ops]
    MuteUserRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
        reason:
          type: string
    UnmuteUserRequest:
      type: object
      required: [user_id]
      properties:
        user_id:
          type: string
          format: uuid
        reason:
          type: string
...
    WalletReconcileResponse:
      type: object
      required: [checked_users, imbalance_count, balanced, mismatches, mismatches_truncated]
      properties:
        checked_users:
          type: integer
          format: int64
        imbalance_count:
          type: integer
          format: int64
        balanced:
          type: boolean
        mismatches:
          type: array
          items:
            $ref: "#/components/schemas/BalanceMismatch"
        mismatches_truncated:
          type: boolean
    BalanceMismatch:
      type: object
      required: [user_id, stored_balance, ledger_sum]
      properties:
        user_id:
          type: string
          format: uuid
        stored_balance:
          type: integer
          format: int64
        ledger_sum:
          type: integer
          format: int64
...
    ExpirePayOrdersResponse:
      type: object
      required: [expired_count]
      properties:
        expired_count:
          type: integer
          format: int64
    AnalyticsNameCount:
      type: object
      required: [name, count]
      properties:
        name: {type: string}
        count: {type: integer, format: int64}
    AnalyticsRecentEvent:
      type: object
      required: [id, user_id, name, occurred_at, received_at]
      properties:
        id: {type: string}
        user_id: {type: string}
        name: {type: string}
        occurred_at: {type: string}
        received_at: {type: string}
    AnalyticsSummaryResponse:
      type: object
      required: [retained_events, distinct_users, by_name, recent]
      properties:
        retained_events: {type: integer, format: int64}
        distinct_users: {type: integer, format: int64}
        by_name:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsNameCount"
        recent:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsRecentEvent"